refactor(rgb-to-color)
